comment out line 138 in label_handling.py

+++ nnunetv2/utilities/label_handling/label_handling.py
@@ -135,7 +135,7 @@ class LabelManager(object):
 
         with torch.no_grad():
             # softmax etc is not implemented for half
-            logits = logits.float()
+            #logits = logits.float()
             probabilities = self.inference_nonlin(logits)
 
         return probabilities
